LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-text-to-the-canvas/null-character.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
</head>
<body class="show_output">
<p class="desc">Testing the NULL character</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing letter spacing and word spacing");
_addTest(function(canvas, ctx) {
    ctx.font = "48px 'Helvetica'";
    let result = ctx.measureText("\u0000");
    _assertSame(result.width, 0, "result.width", "0");
});
</script>
</body>
</html>
